requirements.txt
fastapi
uvicorn
sqlalchemy
inference-sdk
easyocr
opencv-python
python-dotenv
